DCServo device add timeout functionality plus json cfg

+++ src/HAL_JSON/Devices/OneWireTemp/HAL_JSON_OneWireTempBus.cpp
@@ -82,7 +82,7 @@ namespace HAL_JSON {
             if (valid == false) continue;
             deviceCount++;
         }
-        devices = new (std::nothrow) Device*[deviceCount]();
+        devices = new (std::nothrow) Device*[static_cast<size_t>(deviceCount)]();
         uint32_t index = 0;
         for (int i=0;i<itemCount;i++) {
             if (validDevices[i] == false) continue;
